Swap DEFAULT_CONFIG global dict to a function-scoped fixture.
Also add a default_config_with fixture for default plus one change (the
main use case in our parametrised tests).
