+++ tests/test_git_init.py
@@ -5,19 +5,20 @@ import typing
 
 import pytest
 
-from .helpers import DEFAULT_CONFIG  # type: ignore[import-not-found]
-
 
 @pytest.mark.parametrize("initialise_git_repository", [True, False])
 def test_initialisation_of_git_repo(
     initialise_git_repository: bool,  # noqa: FBT001
+    default_config_with: typing.Callable,
     generate_package: typing.Callable,
 ) -> None:
     """Checks to see if git was correctly initialised if desired."""
-    test_config = DEFAULT_CONFIG.copy()
-    test_config["initialise_git_repository"] = str(initialise_git_repository)
+    config = default_config_with(
+        "initialise_git_repository", str(initialise_git_repository)
+    )
+
     # Run cookiecutter with initialise_git_repository
-    result, test_project_dir = generate_package(config=test_config)
+    result, test_project_dir = generate_package(config=config)
 
     # check if git is initialised
     git_status = subprocess.run(  # noqa: S603
@@ -51,9 +52,9 @@ def test_initialisation_of_git_repo(
         )
         assert (
             "GitHub CLI detected, you can create a repo with the following:\n\n"
-            f"gh repo create {test_config['github_owner']}/"
+            f"gh repo create {config['github_owner']}/"
             f"cookiecutter-test -d "
-            f'"{test_config["project_short_description"]}" --public -r '
+            f'"{config["project_short_description"]}" --public -r '
             f"origin --source cookiecutter-test" in result.stdout
         )
     except FileNotFoundError:
@@ -63,11 +64,11 @@ def test_initialisation_of_git_repo(
         assert (
             "You now have a local git repository. To sync this to GitHub you "
             "need to create an empty GitHub repo with the name "
-            f"{test_config['github_owner']}/"
+            f"{config['github_owner']}/"
             f"cookiecutter-test - DO NOT SELECT ANY "
             "OTHER OPTION.\n\nSee this link for more detail "
             "https://docs.github.com/en/get-started/quickstart/create-a-repo"
             ".\n\nThen run:\n\ngit remote add origin git@github.com:"
-            f"{test_config['github_owner']}/"
+            f"{config['github_owner']}/"
             f"cookiecutter-test.git" in result.stdout
         )
